+++ index.html
@@ -3,7 +3,7 @@
         <title>hello</title>
     </head>
     <body>
-        <div id="main"></div>
+        <div id="root"></div>
         <script src="/dist/index.js"></script>
     </body>
 </html>
