+++ src/index.tsx
@@ -1,0 +1,8 @@
+import * as React from "react"
+import * as ReactDom from "react-dom"
+import App from "./app"
+
+ReactDom.render(
+    <App />,
+    document.querySelector('#root')
+)
